+++ Changelog.md
@@ -5,6 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2024-06-10
+
+### Summary
+
+- Work with framework OriCommJS 1.1.1 and above.
+
+### Added
+
+### Changed
+
+- Update example in startup/sqlite. Completed on 2024-06-10
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security
+
+[1.1.0]: https://github.com/wkloh76/oricommjs_components/releases/tag/1.1.0
+
 ## [1.0.9] - 2024-05-27
 
 ### Summary